Pupil sensor parse transform
