+++ PupilInterface/PupilInterface/ZyreNode.cs
@@ -4,7 +4,6 @@ using System.Linq;
 using System.Reactive.Linq;
 using NetMQ.Zyre;
 using NetMQ.Zyre.ZyreEvents;
-using Newtonsoft.Json.Linq;
 using NetMQ;
 
 namespace PupilInterface
