+++ PupilInterface/workflows/test-zyre2.bonsai
@@ -2,6 +2,7 @@
 <WorkflowBuilder Version="2.7.1"
                  xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
                  xmlns:p1="clr-namespace:PupilInterface;assembly=PupilInterface"
+                 xmlns:rx="clr-namespace:Bonsai.Reactive;assembly=Bonsai.Core"
                  xmlns="https://bonsai-rx.org/2018/workflow">
   <Workflow>
     <Nodes>
@@ -11,7 +12,41 @@
           <p1:Group>pupil-mobile-v4</p1:Group>
         </Combinator>
       </Expression>
+      <Expression xsi:type="rx:Condition">
+        <Name>WhisperEvent</Name>
+        <Workflow>
+          <Nodes>
+            <Expression xsi:type="WorkflowInput">
+              <Name>Source1</Name>
+            </Expression>
+            <Expression xsi:type="MemberSelector">
+              <Selector>EventType</Selector>
+            </Expression>
+            <Expression xsi:type="Equal">
+              <Operand xsi:type="StringProperty">
+                <Value>WhisperEvent</Value>
+              </Operand>
+            </Expression>
+            <Expression xsi:type="WorkflowOutput" />
+          </Nodes>
+          <Edges>
+            <Edge From="0" To="1" Label="Source1" />
+            <Edge From="1" To="2" Label="Source1" />
+            <Edge From="2" To="3" Label="Source1" />
+          </Edges>
+        </Workflow>
+      </Expression>
+      <Expression xsi:type="MemberSelector">
+        <Selector>EventType</Selector>
+      </Expression>
+      <Expression xsi:type="MemberSelector">
+        <Selector>Content</Selector>
+      </Expression>
     </Nodes>
-    <Edges />
+    <Edges>
+      <Edge From="0" To="1" Label="Source1" />
+      <Edge From="1" To="2" Label="Source1" />
+      <Edge From="1" To="3" Label="Source1" />
+    </Edges>
   </Workflow>
 </WorkflowBuilder>
